+++ drivers/entropy/entropy_stm32.c
@@ -320,7 +320,7 @@ static int random_byte_get(void)
 			goto out;
 		}
 
-		retval = LL_RNG_ReadRandData32(rng);
+		retval = (int)LL_RNG_ReadRandData32(rng);
 		if (retval == 0) {
 			/* A seed error could have occurred between RNG_SR
 			 * polling and RND_DR output reading.
